+++ src/openapi/streaming/streaming.ts
@@ -66,7 +66,6 @@ type EmittedEvents = {
     [connectionConstants.EVENT_STREAMING_FAILED]: () => void;
     [connectionConstants.EVENT_CONNECTION_SLOW]: () => void;
     [connectionConstants.EVENT_DISCONNECT_REQUESTED]: () => void;
-    [connectionConstants.EVENT_MULTIPLE_ORPHANS_FOUND]: () => void;
     [connectionConstants.EVENT_PROBE_MESSAGE]: (
         message: types.ProbeControlMessage,
     ) => void;
@@ -96,12 +95,6 @@ class Streaming extends MicroEmitter<EmittedEvents> {
      * Event that occurs when server sends _disconnect control message.
      */
     EVENT_DISCONNECT_REQUESTED = connectionConstants.EVENT_DISCONNECT_REQUESTED;
-
-    /**
-     * Event that occurs when we detect a problem with multiple orphans found
-     */
-    EVENT_MULTIPLE_ORPHANS_FOUND =
-        connectionConstants.EVENT_MULTIPLE_ORPHANS_FOUND;
 
     /**
      * Event that occurs when probe message is received.
@@ -446,6 +439,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
 
                 this.shouldSubscribeBeforeStreamingSetup = false;
                 this.orphanFinder.stop();
+                this.orphanEvents = [];
 
                 if (this.isReset) {
                     this.init();
@@ -483,6 +477,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
                 this.updateConnectionQuery();
 
                 this.orphanFinder.stop();
+                this.orphanEvents = [];
                 break;
 
             case this.CONNECTION_STATE_CONNECTED:
@@ -914,8 +909,12 @@ class Streaming extends MicroEmitter<EmittedEvents> {
                 }
             }
 
-            // multiple service paths have failed multiple times, more than 20 seconds apart, within the same minute
-            if (servicePathFailures >= 2) {
+            if (
+                // multiple service paths have failed multiple times, more than 20 seconds apart, within the same minute
+                servicePathFailures >= 2 &&
+                // We haven't received updates for 20 seconds
+                Date.now() - this.latestActivity > 20 * 1000
+            ) {
                 // debugging information...
                 const activities = this.subscriptions.map((sub) => ({
                     latestActivity: sub.latestActivity,
@@ -929,7 +928,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
                 // at the source server
                 log.warn(
                     LOG_AREA,
-                    'Detected multiple service paths hitting orphans, multiple times',
+                    'Detected multiple service paths hitting orphans, multiple times - reconnecting',
                     {
                         now: Date.now(),
                         orphanEvents: this.orphanEvents,
@@ -938,7 +937,9 @@ class Streaming extends MicroEmitter<EmittedEvents> {
                         heartBeatLog: this.heartBeatLog,
                     },
                 );
-                this.trigger(this.EVENT_MULTIPLE_ORPHANS_FOUND);
+
+                // reconnect - we may be disconnected and not know it
+                this.reconnect();
             }
         });
     }
@@ -949,11 +950,16 @@ class Streaming extends MicroEmitter<EmittedEvents> {
      */
     private onOrphanFound(subscription: Subscription) {
         try {
-            log.info(LOG_AREA, 'Subscription has become orphaned - resetting', {
-                referenceId: subscription.referenceId,
-                streamingContextId: subscription.streamingContextId,
-                servicePath: subscription.servicePath,
-            });
+            log.info(
+                LOG_AREA,
+                'Subscription has become orphaned - resetting',
+                {
+                    referenceId: subscription.referenceId,
+                    streamingContextId: subscription.streamingContextId,
+                    servicePath: subscription.servicePath,
+                },
+                { persist: true },
+            );
 
             const now = Date.now();
             // most subscriptions time out after 60 seconds. So we use 70s to give us 10s leniency to detect multiple unsubscribes
@@ -1275,6 +1281,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
      * It *will not* stop the subscription (see dispose for that). It is useful for testing reconnect logic works or for resetting all subscriptions.
      */
     disconnect() {
+        this.orphanEvents = [];
         this.connection.stop();
     }
 
@@ -1288,6 +1295,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         }
 
         this.orphanFinder.stop();
+        this.orphanEvents = [];
 
         for (let i = 0; i < this.subscriptions.length; i++) {
             const subscription = this.subscriptions[i];
@@ -1317,6 +1325,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         this.disposed = true;
 
         this.orphanFinder.stop();
+        this.orphanEvents = [];
 
         for (let i = 0; i < this.subscriptions.length; i++) {
             const subscription = this.subscriptions[i];
@@ -1344,6 +1353,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         this.isReset = true;
 
         this.orphanFinder.stop();
+        this.orphanEvents = [];
 
         if (this.reconnecting) {
             clearTimeout(this.reconnectTimer);